fix(cli): rename session-lifecycle verbs to session-verbs.ts; agent.ts is now the standing-Agent CRUD group (#1417)
The existing agent.ts registered top-level session verbs (spawn/resume/status/
kill/send-input/attach) via registerAgentCommands. Renamed that file to
session-verbs.ts / registerSessionVerbCommands, freeing 'agent' for the new
'grackle agent' entity CRUD group (list/create/show/edit/delete). Resolves the
duplicate-registration crash.
